migrate ErrorDecorator ts
